[aoti-backend-consolidation 2/3] backend.py
Differential Revision: [D85704977](https://our.internmc.facebook.com/intern/diff/D85704977/)

[ghstack-poisoned]
